Changed to latest json.net via nuget, bug fix & .gitignore
Nuget - All the cool kids use it, so I set up the project to use nuget to
get the latest version of json.net

.gitignore - Updated to ignore packages downloaded by nuget

empty responses - In the case of certain kinds of errors the server's
response might not have a response body which causes an exception because
GetJObject requires a response body to exist. Fixed this by adding
a constructor in CouchResponseObject that takes no arguments and
teaching CouchResponse to use that constructor if the response body
has no value.